// src/app/models/forecast.ts
export class Forecast {
  date: string;
  weather: string;
  icon: string;
  clouds: number;
  humidity: number;
  pressure: number;
  windSpeed: number;
  morningTemperature: number;
  dayTemperature: number;
  eveningTemperature: number;
  nightTemperature: number;
  maxTemperature: number;
  minTemperature: number;
  class: string;
}